Population can be set with REST POST/PUT

+++ business/country.go
@@ -8,7 +8,7 @@ import (
 type Country struct {
 	Code       string
 	Name       string
-	Population uint
+	Population uint64
 }
 
 type CountryList []Country
